refactor: split configuration groups
