+++ src/sparsevllm/configs/__init__.py
@@ -1,0 +1,21 @@
+"""Structured Sparse-vLLM configuration components."""
+
+from sparsevllm.configs.groups import (
+    DecodeCudaGraphConfig,
+    DeltaKVConfig,
+    ObservabilityConfig,
+    PrefixCacheConfig,
+    SparseMethodConfig,
+)
+from sparsevllm.configs.runtime import Config, QuantizationConfig, RuntimeLayout
+
+__all__ = [
+    "Config",
+    "DecodeCudaGraphConfig",
+    "DeltaKVConfig",
+    "ObservabilityConfig",
+    "PrefixCacheConfig",
+    "QuantizationConfig",
+    "RuntimeLayout",
+    "SparseMethodConfig",
+]
