+++ tests/test_decode_cuda_graph_tp_config.py
@@ -27,7 +27,7 @@ class DecodeCudaGraphTPConfigTest(unittest.TestCase):
         with tempfile.TemporaryDirectory() as tmp:
             model_dir = Path(tmp) / model_name
             model_dir.mkdir()
-            with patch("sparsevllm.config.AutoConfig.from_pretrained", return_value=self.hf_config()):
+            with patch("sparsevllm.configs.runtime.AutoConfig.from_pretrained", return_value=self.hf_config()):
                 return Config(
                     model=str(model_dir),
                     vllm_sparse_method=method,
