+++ tests/test_research_fail_fast.py
@@ -476,7 +476,7 @@ class ResearchFailFastTest(unittest.TestCase):
                 '{"model_type": "qwen2", "torch_dtype": "float16", "max_position_embeddings": 32768}\n',
                 encoding="utf-8",
             )
-            with patch("sparsevllm.config.AutoConfig.from_pretrained", side_effect=RuntimeError("boom")):
+            with patch("sparsevllm.configs.runtime.AutoConfig.from_pretrained", side_effect=RuntimeError("boom")):
                 with self.assertRaisesRegex(RuntimeError, "Refusing to silently fall back"):
                     Config(model=str(model_dir))
 
@@ -490,7 +490,7 @@ class ResearchFailFastTest(unittest.TestCase):
             num_hidden_layers=2,
         )
         with tempfile.TemporaryDirectory() as tmp:
-            with patch("sparsevllm.config.AutoConfig.from_pretrained", return_value=hf_config):
+            with patch("sparsevllm.configs.runtime.AutoConfig.from_pretrained", return_value=hf_config):
                 with self.assertRaisesRegex(NotImplementedError, "Unsupported Sparse-vLLM model_type"):
                     Config(model=tmp)
 
@@ -504,7 +504,7 @@ class ResearchFailFastTest(unittest.TestCase):
             num_hidden_layers=2,
         )
         with tempfile.TemporaryDirectory() as tmp:
-            with patch("sparsevllm.config.AutoConfig.from_pretrained", return_value=hf_config):
+            with patch("sparsevllm.configs.runtime.AutoConfig.from_pretrained", return_value=hf_config):
                 with self.assertRaisesRegex(ValueError, "requires deltakv_path"):
                     Config(model=tmp, vllm_sparse_method="deltakv")
                 with self.assertRaisesRegex(ValueError, "requires deltakv_path"):
